requirements.txt
# Backend dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright>=1.39.0
pillow==11.1.0
reportlab>=4.0.7
python-multipart==0.0.6
pymongo>=4.6.0
anthropic>=0.8.0
openai>=1.3.5
tqdm==4.66.1
loguru==0.7.2

# For testing
pytest>=7.4.3
httpx>=0.25.1